requirements.txt
Backtesting==0.3.3
numpy==1.20.3
pandas==1.3.4
pandas_ta==0.3.14b0
Pillow==9.0.0
pyTelegramBotAPI==4.3.1
requests==2.26.0
tinydb==4.6.1
yfinance==0.1.67
